k256/p256: factor out `affine` and `projective` submodules
Refactors the `arithmetic` module, extracting `affine` and `projective`
modules which now contain `AffinePoint` and `ProjectivePoint`
respectively.
