+++ k256/bench/bench.rs
@@ -2,23 +2,29 @@ use core::convert::TryInto;
 use criterion::measurement::Measurement;
 use criterion::{criterion_group, criterion_main, BenchmarkGroup, Criterion};
 
-use k256::{FieldElement, ProjectivePoint, Scalar};
+use k256::{elliptic_curve::FromBytes, FieldElement, ProjectivePoint, Scalar};
 
 fn test_scalar_x() -> Scalar {
-    Scalar::from_bytes(&[
-        0xbb, 0x48, 0x8a, 0xef, 0x41, 0x6a, 0x41, 0xd7, 0x68, 0x0d, 0x1c, 0xf0, 0x1d, 0x70, 0xf5,
-        0x9b, 0x60, 0xd7, 0xf5, 0xf7, 0x7e, 0x30, 0xe7, 0x8b, 0x8b, 0xf9, 0xd2, 0xd8, 0x82, 0xf1,
-        0x56, 0xa6,
-    ])
+    Scalar::from_bytes(
+        &[
+            0xbb, 0x48, 0x8a, 0xef, 0x41, 0x6a, 0x41, 0xd7, 0x68, 0x0d, 0x1c, 0xf0, 0x1d, 0x70,
+            0xf5, 0x9b, 0x60, 0xd7, 0xf5, 0xf7, 0x7e, 0x30, 0xe7, 0x8b, 0x8b, 0xf9, 0xd2, 0xd8,
+            0x82, 0xf1, 0x56, 0xa6,
+        ]
+        .into(),
+    )
     .unwrap()
 }
 
 fn test_scalar_y() -> Scalar {
-    Scalar::from_bytes(&[
-        0x67, 0xe2, 0xf6, 0x80, 0x71, 0xed, 0x82, 0x81, 0xe8, 0xae, 0xd6, 0xbc, 0xf1, 0xc5, 0x20,
-        0x7c, 0x5e, 0x63, 0x37, 0x22, 0xd9, 0x20, 0xaf, 0xd6, 0xae, 0x22, 0xd0, 0x6e, 0xeb, 0x80,
-        0x35, 0xe3,
-    ])
+    Scalar::from_bytes(
+        &[
+            0x67, 0xe2, 0xf6, 0x80, 0x71, 0xed, 0x82, 0x81, 0xe8, 0xae, 0xd6, 0xbc, 0xf1, 0xc5,
+            0x20, 0x7c, 0x5e, 0x63, 0x37, 0x22, 0xd9, 0x20, 0xaf, 0xd6, 0xae, 0x22, 0xd0, 0x6e,
+            0xeb, 0x80, 0x35, 0xe3,
+        ]
+        .into(),
+    )
     .unwrap()
 }
